Add `BlobStore.deflateContent`.
